feat: Read Data Source, Schema & Resolver configuration from posgres
The `sequelize` module is used for this.
A `DataSource`, `GraphQLSchema` & `Resolver` model is defined.
On startup, the first schema is retrieved from the database,
and all Resolvers & DataSources are read and passed into the schema parser
logic.

The local dev setup has changed due to the introduction of postgres.
See README for details.
The gist of it is:

```
npm run db:start
npm run db:seed      # Creates sequelize model tables & seeds some example data
npm run dev
```

to stop:

```
npm run db:stop
```

The server is expected to fail starting at this time as the `InMemory`
data source is not implemented yet.
You should see an error like this on startup if the db is working OK.

```
Error: Unhandled data source type: InMemory
    at _.forEach (/home/dmartin/work/data-sync-server/server/lib/dataSourceParser.js:16:13)
    at arrayEach (/home/dmartin/work/data-sync-server/node_modules/lodash/lodash.js:516:11)
    at Function.forEach (/home/dmartin/work/data-sync-server/node_modules/lodash/lodash.js:9342:14)
    at module.exports (/home/dmartin/work/data-sync-server/server/lib/dataSourceParser.js:7:5)
    at module.exports (/home/dmartin/work/data-sync-server/server/lib/schemaParser.js:7:23)
    at module.exports (/home/dmartin/work/data-sync-server/server/server.js:32:14)
    at <anonymous>
```

This error should not happen after the `InMemory` data source and
resolver mappings have been added to the server.